Change WorksForm from products app to admininstration
